+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## V1.2.0
+
+### Added
+- Filters on package list table
+- Package suggestions on creation view
+
+ ### Modified
+ - Update dependencies
+
 ## v1.1.0
 
 ### Added
@@ -8,7 +17,7 @@
 - New "Details" view
 - Handle private registry in addition to the public registry
 
-### Modify
+### Modified
 - Application won't open a window at start-up
 - Better error management
 - Update dependencies